Improve performance by using native array
Since the implementation already has to update and read from the
cursors on each set or read operation, there is no need to track
individual access to items *within* the array. Accordingly, drop
`TrackedArray` and just use a native array as the backing storage.

This does unfortunately mean that every update *will* cause a rerender,
but:

- that was already true!
- it has the nice benefit of not also creating additional tags for
  every access of the items in the array.