Remove extraneous bindings and Cursor class
None of these cost very much, but there's no reason to carry them
around at all: the extra `Cursors` class was a legacy of the original
implementation which was trying to model the system as a tagged union
type, and the cursors can simply live directly on the queue.
Additionally, dropping any intermediate bindings which aren't needed
cuts down juuuuust a tiny bit of weight in the final implementation.