Add tests for autotracking behavior
While integrating this into an app, we discovered that there is a case
where it is possible to trigger a rerender assertion: any time that in
the course of rendering, you access a queue instance you have already
read. This is actually reproducible with `TrackedQueue.map()`, since
the currently implementation *always* reads both `_head` and `_tail`
when performing any operations, since those are used as the basis for
computing the current size of the queue, which in turn is necessary for
computing the wrapping behavior of the queue.

Here, simply introduce sufficient coverage to (a) show that bug and (b)
make sure the changes made to fix that bug do not cause regressions in
the behavior of the rest of the system.