Separate reads from writes in tracked data
All write operations should be possible to execute without ever reading
from tracked data. Otherwise, end users can accidentally end up with
unresolveable bugs where they are, from their perspective, only writing
to tracked state, but the library itself is *reading* its own tracked
state internally as part of the process of writing the new values into
the storage.

To enable this, switch to simply dirtying the collection itself on all
write operations, and consuming it only in read operations. This
ultimately has the same semantics as the previous, because:

- we always had to invalidate the whole collection every time either
  cursor was written, since iteration (via `Symbol.iterator`) would need
  to be re-executed regardless of *what* the mutation operation was
- all read operations always read through `size`, which necessarily had
  to read both cursors

To accomplish this, however, we cannot (yet!) simply create a tracked
property which represents the collection and set it in the case of a
change to the collection, because we support Ember 3.16 and Ember 3.20,
where iteration (via `Symbol.iterator` or the `forEach` fallback) is
not autotracked, and therefore needs special handling. Instead,
introduce two new functions in the module, `consumeCollection` and
`dirtyCollection`, which do exactly what their names imply, but with the
important property of doing it *differently* for versions prior to Ember
3.22, which solved this problem by properly autotracking everything in
the system, including iteration.

Here, use the Embroider macro system to fall back to the classic Ember
`get` and `notifyPropertyChange` semantics for those earlier versions,
but use the tracked storage primitives (via polyfill) for versions where
that issues is resolved. In the future, once we drop support for earlier
versions of Ember, this can be simplified to define the functions as
normal functions in module scope *without* the macro definition.

It will still remain the case that simply dirtying and consuming the
collection itself will be the right design here (with lower overhead
than tracking the cursors), but it may at that point make sense to
simply have a method on the class itself along with a per-instance
tracked storage which it interacts with, thereby eliminating the need
for the `WeakMap` and storage management. (That could even be a simple
tracked property, but the direct use of storage is likely actually
*clearer* in that case.)