Disable CI runs for ember stable/beta/canary
Since stable, beta, and canary are now in Ember v4, and supporting
Ember v4 requires making a breaking release to this library (by way of
requiring ember-auto-import v2 and webpack v5 as actual dependencies),
drop them from the CI matrix for now. They should be added back in when
adding support for Ember 4.