Upgrade release-it to latest and fix vm2 vulnerability
The vulnerability *only* exists in the release flow, and this makes
sure that we are using a version of release-it which transitively gets
what should be a compatible version of `vm2`, and then updates the lock
file to resolve at least the version of `vm2` with the fix (3.9.15). We
are now safe, since Yarn will never *regress* the resolved lock file
version.